feat: add negative test fixtures to Rust
Also run the negative test fixtures on the Rust implementations.

This commit also upgrades to the newest version of libipld, which
now passes all DAG-PB fixtures, except for the integer out of i64
range one.

It also refactors the code a bit, so that the libipld codecs and
the Serde based ones share more of the code.

Supersedes #78.